requirements.txt
python>=3.8
requests>=2.29.0
openpyxl>=2.4.0
build>=0.9.0
twine>=3.8.0
pycryptodome>=3.18.0
fpdf>=1.7.2
PyPDF2>=3.0.1
Pillow>=9.4.0
setuptools>=67.8.0
beautifulsoup4>=4.12.2
PyMuPDF>=1.23.0
python-docx>=1.1.0